APIenv/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-rest-framework==0.1.0
djangorestframework==3.15.1
idna==3.7
requests==2.32.2
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
